Перезаливаю исправленный код в новую ветку. Модели вынесены в самостоятельные файлы, Лишние методы удалены, все исправлено для четкого соответсвия с требованиями заданий.
