+++ BankSystem.App/Services/BankService.cs
@@ -1,16 +1,15 @@
-﻿using System;
+﻿using BankSystemDomen.Modelss;
+using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static BankSystem.Class;
+
+using static BankSystem.Person;
 
 namespace BankSystem.App.Services
 {
     public class BankService
-    {
-        // ( ** ) Метод для введения данных о расходах и доходах
-        public List<int> ProfitAndCosts()
+    {    
+        // ( ** ) Метод для расчета зарплаты акционера
+        public double SalaryBankShareholders() 
         {
             Console.WriteLine("Введите предполагаемые доходы: ");
             string i = Console.ReadLine();
@@ -20,25 +19,18 @@ namespace BankSystem.App.Services
             string j = Console.ReadLine();
             int J = int.Parse(j);
 
-            List<int> ProfitCosts = new List<int> { I,J };
+            List<int> ProfitCosts = new List<int> { I, J };
             ProfitCosts[0] = I;
             ProfitCosts[1] = J;
-            return ProfitCosts;
-        }
-        // ( ** ) Метод для выведения предполагаемого количества акционеров банка
-        public int AmountBankShareholders()
-        {
+
             Console.WriteLine("Введите количество акционеров банка: ");
             string t = Console.ReadLine();
-            int T = int.Parse(t);
-            return T;
-        }
-        // ( ** ) Метод для расчета зарплаты акционера
-        public double SalaryBankShareholders(List<int> profitandcosts, int amount) 
-        {
-            double salary = (profitandcosts[0] - profitandcosts[1])/amount;
+            int amount = int.Parse(t);
+
+            double salary = (ProfitCosts[0] - ProfitCosts[1])/amount;
             return salary;
         }
+         
         // ( ** ) Метод для преобразования клиента банка в сотрудника
         public Employee Recruitment(Client client)
         {
@@ -47,15 +39,7 @@ namespace BankSystem.App.Services
             employee.Salary = 1000;
             employee.Post = "Разработчик"; 
             return employee;
-
         }
-        // ( ** ) Метод для добавления и обновления значения свойства Contract экземпляра класса Employee
-        public string UpgradeContract(Employee employee)
-        {
-            Console.WriteLine("Введите наименование нового контракта: ");
-            string offer = Console.ReadLine();
-            employee.Contract = offer;
-            return employee.Contract;
-        }
+        
     }
 }
