+++ BankSystemApplication/Modelss/Currency.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BankSystemDomen.Modelss
+{
+    public struct Currency
+    {
+        public string Name { get; set; }
+        public int Id { get; set; }
+        public string Country { get; set; }
+    }
+}
