+++ BankSystemApplication/Modelss/Employee.cs
@@ -1,0 +1,16 @@
+﻿using BankSystem;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BankSystemDomen.Modelss
+{
+    public class Employee  :  Person
+    {
+        public string Contract { get; set; }
+        public double Salary { get; set; }
+        public string Post { get; set; }
+    }
+}
