+++ BankSystemApplication/Modelss/Person.cs
@@ -1,0 +1,17 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BankSystem
+{
+    public class Person
+    {
+        public int id { get; set; }
+        public string name { get; set; }
+
+    }
+
+}
